requirements.txt
python-uds==0.3.2
python-can==4.3.1
cantools==39.3.0
